refactor(ui): add nav contacts component
