+++ app/shared/components/design-system/all-components/navigation-bar/mobile-nav/mobile-overlay/MobileMenuOverlay.tsx
@@ -1,18 +1,13 @@
-import { Box, Slide, Typography } from '@mui/material';
-import { useTranslations } from 'next-intl';
+import { Box, Slide } from '@mui/material';
 import { useEffect, useMemo } from 'react';
 
-import LanguageSwitcher from '../../../language-switcher/LanguageSwitcher';
 import { NavAccordion } from '../../../menu-title/NavAccordion';
+import { ContactsSection } from '../nav-contacts-section/NavContacts';
 import { styles } from './MobileNavOverlay.styles';
 import { contactsData, LinkIcon } from '~/types/types/common.types';
 
 import { NavigationDTO } from '~/domain/dto/navigation.dto';
-import MailIcon from '~/public/icons/mail.svg';
-import PhoneIcon from '~/public/icons/phone.svg';
 import { ColumnGuides } from '~/shared/components/column-guides/ColumnGuides';
-import { ContactLink } from '~/shared/components/contact-link/ContactLink';
-import FooterSocialMedia from '~/shared/components/Footer/footer-social-media/FooterSocialMedia';
 import useBreakpoints from '~/shared/hooks/use-breakpoints/useBreakpoints';
 
 interface MobileMenuOverlayProps {
@@ -23,7 +18,6 @@ interface MobileMenuOverlayProps {
 }
 
 const MobileMenuOverlay = ({ open, navLabels, contacts, socialLinks }: MobileMenuOverlayProps) => {
-  const t = useTranslations('contactsInfoPage');
   const { isMobile } = useBreakpoints();
 
   useEffect(() => {
@@ -53,38 +47,16 @@ const MobileMenuOverlay = ({ open, navLabels, contacts, socialLinks }: MobileMen
       <Box sx={styles.overlay}>
         <ColumnGuides lineColor="rgba(239, 233, 224, 0.3)" />
 
-        <Box sx={styles.leftColumn}>
-          {isMobile && <LanguageSwitcher variant="mobile" />}
-          <Box sx={styles.links}>
-            <ContactLink
-              type="phone"
-              label={isMobile ? undefined : t('phoneNumber')}
-              value={contacts.phone}
-              isMobile={isMobile}
-              direction={isMobile ? 'row' : 'column'}
-              data-testid="ContactsInfo-phoneLink"
-              icon={isMobile ? PhoneIcon : null}
-            />
-            <ContactLink
-              type="email"
-              label={isMobile ? undefined : t('email')}
-              value={contacts.email}
-              isMobile={isMobile}
-              direction={isMobile ? 'row' : 'column'}
-              data-testid="ContactsInfo-emailLink"
-              icon={isMobile ? MailIcon : null}
-            />
+        {!isMobile ? (
+          <Box sx={styles.leftColumn}>
+            <ContactsSection contacts={contacts} socialLinks={socialLinks} isMobile={isMobile} />
           </Box>
-          <Box sx={{ display: 'flex', flexDirection: 'column', gap: { sm: '8px', md: '10px' } }}>
-            <Typography sx={styles.mediaTitles} data-testid="ContactsInfo-socialMediaLabel">
-              {t('socialMedia')}:
-            </Typography>
-
-            <FooterSocialMedia media={socialLinks} containerSx={{ gap: { xs: '18px', sm: '14px', md: '16px' } }} />
-          </Box>
-        </Box>
+        ) : null}
         <Box sx={styles.rightColumn}>
           <NavAccordion items={navItems} />
+          <Box sx={styles.contactsContainer}>
+            {isMobile && <ContactsSection contacts={contacts} socialLinks={socialLinks} isMobile={isMobile} />}
+          </Box>
         </Box>
       </Box>
     </Slide>
