+++ app/shared/components/design-system/all-components/navigation-bar/mobile-nav/nav-contacts-section/NavContacts.tsx
@@ -1,0 +1,51 @@
+import { Box, Typography } from '@mui/material';
+import { useTranslations } from 'next-intl';
+
+import LanguageSwitcher from '../../../language-switcher/LanguageSwitcher';
+import { styles } from './NavContacts.styles';
+import { contactsData, LinkIcon } from '~/types/types/common.types';
+
+import MailIcon from '~/public/icons/mail.svg';
+import PhoneIcon from '~/public/icons/phone.svg';
+import { ContactLink } from '~/shared/components/contact-link/ContactLink';
+import FooterSocialMedia from '~/shared/components/Footer/footer-social-media/FooterSocialMedia';
+
+interface ContactsSectionProps {
+  contacts: contactsData;
+  socialLinks: LinkIcon[];
+  isMobile: boolean;
+}
+
+export const ContactsSection = ({ contacts, socialLinks, isMobile }: ContactsSectionProps) => {
+  const t = useTranslations('contactsInfoPage');
+
+  return (
+    <>
+      {isMobile && <LanguageSwitcher variant="mobile" />}
+
+      <Box sx={styles.links}>
+        <ContactLink
+          type="phone"
+          label={isMobile ? undefined : t('phoneNumber')}
+          value={contacts.phone}
+          isMobile={isMobile}
+          direction={isMobile ? 'row' : 'column'}
+          icon={isMobile ? PhoneIcon : null}
+        />
+        <ContactLink
+          type="email"
+          label={isMobile ? undefined : t('email')}
+          value={contacts.email}
+          isMobile={isMobile}
+          direction={isMobile ? 'row' : 'column'}
+          icon={isMobile ? MailIcon : null}
+        />
+      </Box>
+
+      <Box sx={styles.socialMediaBox}>
+        <Typography sx={styles.mediaTitles}>{t('socialMedia')}:</Typography>
+        <FooterSocialMedia media={socialLinks} containerSx={{ gap: { xs: '18px', sm: '14px', md: '16px' } }} />
+      </Box>
+    </>
+  );
+};
